requirements.txt
pyvis
treelib
pyyaml
torch
InstructorEmbedding
sentence_transformers